requirements.txt
# Email Integration
imapclient==3.0.1
email-validator==2.1.0

# File Processing
python-docx==1.1.0
PyPDF2==3.0.1
openpyxl==3.1.2

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Configuration
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dateutil==2.8.2

# NLP and Text Processing
spacy==3.7.2
scikit-learn==1.3.2
nltk==3.8.1

# Matching & Embeddings (optional)
# Updated to avoid cached_download issues with newer huggingface-hub
sentence-transformers==3.0.1  # For semantic matching
transformers==4.44.2
huggingface-hub>=0.34.0,<1.0
openai==1.40.2  # For OpenAI LLM matching (optional; compatible with langchain-openai)
google-generativeai==0.3.2  # For Google Gemini via AI Studio (optional)
faiss-cpu==1.7.4  # For FAISS vector index (optional)

# FastAPI and Web Server
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6  # For file uploads

# Frontend and client
streamlit==1.37.0
requests==2.31.0

# Optional orchestration for LLM (LangChain)
langchain==0.2.10
langchain-core==0.2.35
langchain-openai==0.1.20
